Update livestock_exports.py
Added compute_system_total_protein_exports function for calculation of total protein coming from milk and beef, using 23% for beef and 3.5% for milk.
Changed compute_system_protien_exports methodolgy to make sure that calculation is done correctly, based on numbers of cattle in 1-2yr old cohorts and making sure that LWG from all age groups are accounted for. The efficiency is still not optimised, I struggled with this using ChatGPT for quite a bit and couldn't quite get it over the edge. The structure was similar previously, so not sure how much room for improvement there is.