Update livestock_exports.py
Changed the approach slightly to make it similar to the calculation of weights in the livestock.py file. 
Instead of deleting the previously unused EXPORT_WEIGHT_KEYS in the livestock_data_manager, I changed it into a lookup dictionary that stores the cohort needed for population (steers/heifers between 1-2 years) and the live weight gain cohorts for each age group needed for respective beef output calculation. Beef output results are the same as in my previous version, code is reduced by about 2/3.